fix: basic automation and fix tests path
