+++ maestro-cli/src/test/resources/location/assert_multiple_locations.yaml
@@ -1,0 +1,14 @@
+appId: "com.google.android.apps.maps"
+---
+- launchApp
+- tapOn:
+    text: Skip
+    optional: true
+- setLocation:
+    longitude: 2.295188
+    latitude: 48.8578065
+- assertVisible: .*Eiffel.*
+- setLocation:
+    latitude: 43.7230
+    longitude: 10.3966
+- assertVisible: .*(Piazza|Pisa).*
