+++ maestro-cli/src/test/resources/travel/assert_travel_command.yaml
@@ -1,0 +1,25 @@
+appId: "com.google.android.apps.maps"
+---
+- launchApp
+- tapOn:
+    text: Skip
+    optional: true
+- setLocation:
+    latitude: 48.8578065
+    longitude: 2.295188
+- assertVisible: .*Eiffel.*
+- travel:
+    points:
+      - 48.8578065, 2.295188
+      - 46.2276, 5.9900
+      - 43.7230, 10.3966
+      - 41.8902, 12.4922
+    speed: 1000
+- assertVisible: .*Colosseo.*
+
+
+
+
+
+
+
