--- maestro-orchestra/src/test/java/maestro/orchestra/android/AndroidDeviceProvider.kt
@@ -1,13 +1,0 @@
-package maestro.orchestra.android
-
-import dadb.Dadb
-import dadb.adbserver.AdbServer
-
-class AndroidDeviceProvider {
-
-    fun local(): Dadb {
-        val dadb = AdbServer.createDadb(connectTimeout = 60_000, socketTimeout = 60_000)
-
-        return dadb
-    }
-}
