--- maestro-orchestra/src/test/resources/location/assert_multiple_locations.yaml
@@ -1,11 +1,0 @@
-appId: "com.google.android.apps.maps"
----
-- launchApp
-- setLocation:
-    longitude: 2.295188
-    latitude: 48.8578065
-- assertVisible: .*Eiffel.*
-- setLocation:
-    latitude: 43.7230
-    longitude: 10.3966
-- assertVisible: .*Piazza.*
